+++ src/schema/FormSchema.stories.tsx
@@ -3,6 +3,7 @@ import { FormProvider } from 'react-hook-form'
 import { DevTool } from '@hookform/devtools'
 import type { Meta, StoryFn } from '@storybook/react'
 
+import { BottomSheet } from '@/components/bottomSheet/BottomSheet'
 import { Button } from '@/components/button/Button'
 import { InputGroup } from '@/components/inputGroup'
 import Label from '@/components/label/Label'
@@ -10,12 +11,30 @@ import Label from '@/components/label/Label'
 import { useAccountEditForm } from './useAccountEditForm'
 import { useClinicForm } from './useClinicForm'
 import { useLoginForm } from './useLoginForm'
-import { useMedicineForm } from './useMedicineForm'
-import { usePasswordEditForm } from './usePasswordEditForm'
 import { useSignupForm } from './useSignupForm'
 
 const Container = ({ children }: PropsWithChildren) => {
   return <div className="max-w-[450px]">{children}</div>
+}
+
+type TestBottomSheetType = {
+  isShowing: boolean
+  onClickScrim: () => void
+  onClickButton: () => void
+}
+
+const TestBottomSheet = ({ onClickScrim, isShowing, onClickButton }: TestBottomSheetType) => {
+  const handleClickButton = () => {
+    onClickScrim()
+    onClickButton()
+  }
+  return (
+    <BottomSheet onClickScrim={onClickScrim} isShowing={isShowing}>
+      <Button type="button" onClick={handleClickButton}>
+        입력 완료
+      </Button>
+    </BottomSheet>
+  )
 }
 
 export default {
@@ -177,68 +196,6 @@ export const ClinicForm: StoryFn = () => {
   )
 }
 
-export const MedicineForm: StoryFn = () => {
-  const { formMethod, handleSubmitMedicineForm } = useMedicineForm()
-  const { handleSubmit, control } = formMethod
-
-  return (
-    <Container>
-      <FormProvider {...formMethod}>
-        <form onSubmit={handleSubmit(handleSubmitMedicineForm)}>
-          <InputGroup>
-            <Label>약품명</Label>
-            <InputGroup.Input section="medicineName" placeholder="약품명을 입력해주세요." />
-            <InputGroup.ErrorMessage section="medicineName" />
-          </InputGroup>
-
-          <InputGroup direction="row">
-            <Label>복용 요일</Label>
-            <InputGroup.TextWithArrow section="medicineWeek" onClick={() => {}} />
-            <InputGroup.ErrorMessage section="medicineWeek" />
-          </InputGroup>
-
-          <InputGroup direction="row">
-            <Label>복용 횟수</Label>
-            <InputGroup.Stepper section="medicineTimes" />
-          </InputGroup>
-
-          <InputGroup direction="row">
-            <Label>복용 시간</Label>
-            <InputGroup.TextWithArrow section="medicineWeek" onClick={() => {}} />
-          </InputGroup>
-
-          <InputGroup direction="row">
-            <Label>복용 기간</Label>
-            <InputGroup.TextWithArrow section="medicineWeek" onClick={() => {}} />
-            <InputGroup.ErrorMessage section="phoneNumber" />
-          </InputGroup>
-
-          <InputGroup direction="row">
-            <Label>복용량</Label>
-            <InputGroup.TextWithArrow section="medicineWeek" onClick={() => {}} />
-            <InputGroup.ErrorMessage section="medicineWeek" />
-          </InputGroup>
-
-          <InputGroup>
-            <Label>메모</Label>
-            <InputGroup.TextArea section="medicalMemo" placeholder="메모를 기록해주세요." />
-          </InputGroup>
-
-          <InputGroup direction="row">
-            <Label>알람</Label>
-            <InputGroup.TextWithArrow section="medicalAlarm" onClick={() => {}} />
-          </InputGroup>
-
-          <Button primary type="submit" onClick={() => {}}>
-            회원가입
-          </Button>
-        </form>
-      </FormProvider>
-      <DevTool control={control}></DevTool>
-    </Container>
-  )
-}
-
 export const AccountEditForm: StoryFn = () => {
   const { formMethod, handleSubmitAccountEditForm } = useAccountEditForm()
   const { handleSubmit, control } = formMethod
@@ -269,54 +226,6 @@ export const AccountEditForm: StoryFn = () => {
             <Label>성별</Label>
             <InputGroup.Gender />
             <InputGroup.ErrorMessage section="gender" />
-          </InputGroup>
-
-          <Button primary type="submit" onClick={() => {}}>
-            완료
-          </Button>
-        </form>
-      </FormProvider>
-      <DevTool control={control}></DevTool>
-    </Container>
-  )
-}
-
-export const PasswordEditForm: StoryFn = () => {
-  const { formMethod, handleSubmitPasswordEditForm } = usePasswordEditForm()
-  const { handleSubmit, control } = formMethod
-
-  return (
-    <Container>
-      <FormProvider {...formMethod}>
-        <form onSubmit={handleSubmit(handleSubmitPasswordEditForm)}>
-          <InputGroup>
-            <Label>새 비밀번호</Label>
-            <InputGroup.Input
-              section="newPassword"
-              placeholder="8~16자/영문자, 숫자 모두 혼용"
-              type="password"
-            />
-            <InputGroup.ErrorMessage section="newPassword" />
-          </InputGroup>
-
-          <InputGroup>
-            <Label>새 비밀번호 확인</Label>
-            <InputGroup.Input
-              section="confirm"
-              placeholder="비밀번호를 다시 입력해주세요."
-              type="password"
-            />
-            <InputGroup.ErrorMessage section="confirm" />
-          </InputGroup>
-
-          <InputGroup>
-            <Label>현재 비밀번호</Label>
-            <InputGroup.Input
-              section="password"
-              placeholder="기존의 비밀번호를 입력해주세요."
-              type="password"
-            />
-            <InputGroup.ErrorMessage section="password" />
           </InputGroup>
 
           <Button primary type="submit" onClick={() => {}}>
